PaymentIntentSucceededWebHook refactored.
